scripts: aospgsidebloat
Desired users: Samsung owners.
Occasionally, a GSI will be flashed but the cost of a generic system results in a heap of bloat. It is a habit of users to debloat all the unnecessary files, which becomes a tedious process and sometimes a user may get too complacent and delete something they shouldn't have. Therefore, a more safe, predefined approach would be to script what should be debloated.