scripts: make: Stream terminal output to log file
Often when building in GCloud SSH, when streaming the later outputs, the previous gets wiped. If a build error were to occur, this would result in lines being missed particularly lines containing potential build errors. Therefore, if we save the outputs as they occur to a log file we can refer to them for a more reliable understanding of the errors and its context.